+++ modules/broker-core/pom.xml
@@ -117,6 +117,11 @@
             <groupId>com.fasterxml.jackson.core</groupId>
             <artifactId>jackson-annotations</artifactId>
         </dependency>
+        <dependency>
+            <groupId>io.ballerina.messaging</groupId>
+            <artifactId>broker-core</artifactId>
+            <version>0.970.6-SNAPSHOT</version>
+        </dependency>
     </dependencies>
 
     <build>
